requirements.txt
# Core frameworks
fastapi>=0.104.0
streamlit>=1.28.0
uvicorn[standard]>=0.24.0

# LLM and AI
litellm>=1.40.0
anthropic>=0.7.0
google-generativeai>=0.3.0
openai>=1.3.0

# Vector store and embeddings
qdrant-client>=1.6.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Reranking
transformers>=4.30.0
torch>=2.0.0

# Web crawling
playwright==1.40.0
trafilatura==1.6.4
unstructured[html]==0.11.6
beautifulsoup4==4.12.2
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0

# Database and caching
redis==5.0.1
psycopg2-binary==2.9.9  # for PostgreSQL if needed

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
jinja2==3.1.2
httpx==0.25.2

# Observability
langfuse==2.21.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1